// api/model/authTokens.ts
/**
 * Ecommerce Demo API
 * Stable, codegen-oriented OpenAPI contract. Designed for multi-project, multi-language use (client/server stubs, DTOs, SDKs).
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AuthTokens {
    /**
     * Access JWT
     */
    'token': string;
    /**
     * Refresh token if returned by backend
     */
    'refreshToken'?: string;
    /**
     * Access token expiry in seconds
     */
    'expiresIn'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'token',
            baseName: 'token',
            type: 'string'
        },
        {
            name: 'refreshToken',
            baseName: 'refreshToken',
            type: 'string'
        },
        {
            name: 'expiresIn',
            baseName: 'expiresIn',
            type: 'number'
        }
    ];

    static getAttributeTypeMap() {
        return AuthTokens.attributeTypeMap;
    }
}