// api/model/cartItem.ts
/**
 * Ecommerce Demo API
 * Stable, codegen-oriented OpenAPI contract. Designed for multi-project, multi-language use (client/server stubs, DTOs, SDKs).
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CartItem {
    /**
     * Resource identifier
     */
    'productId': string;
    'quantity': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'productId',
            baseName: 'productId',
            type: 'string'
        },
        {
            name: 'quantity',
            baseName: 'quantity',
            type: 'number'
        }
    ];

    static getAttributeTypeMap() {
        return CartItem.attributeTypeMap;
    }
}