// api/model/cartResponse.ts
/**
 * Ecommerce Demo API
 * Stable, codegen-oriented OpenAPI contract. Designed for multi-project, multi-language use (client/server stubs, DTOs, SDKs).
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CartItem } from './cartItem';
import { CartSummaryResponse } from './cartSummaryResponse';

export class CartResponse {
    'items': Array<CartItem>;
    'summary': CartSummaryResponse;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'items',
            baseName: 'items',
            type: 'Array<CartItem>'
        },
        {
            name: 'summary',
            baseName: 'summary',
            type: 'CartSummaryResponse'
        }
    ];

    static getAttributeTypeMap() {
        return CartResponse.attributeTypeMap;
    }
}