// api/model/cartSummaryResponse.ts
/**
 * Ecommerce Demo API
 * Stable, codegen-oriented OpenAPI contract. Designed for multi-project, multi-language use (client/server stubs, DTOs, SDKs).
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CartSummaryResponse {
    /**
     * Number of distinct cart lines/items
     */
    'itemsCount': number;
    /**
     * Sum of quantities across all items
     */
    'totalQuantity': number;
    /**
     * Sum of item prices * quantity (before tax/shipping/discounts)
     */
    'total': number;
    /**
     * ISO-4217 currency code (e.g. USD)
     */
    'currency'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'itemsCount',
            baseName: 'itemsCount',
            type: 'number'
        },
        {
            name: 'totalQuantity',
            baseName: 'totalQuantity',
            type: 'number'
        },
        {
            name: 'total',
            baseName: 'total',
            type: 'number'
        },
        {
            name: 'currency',
            baseName: 'currency',
            type: 'string'
        }
    ];

    static getAttributeTypeMap() {
        return CartSummaryResponse.attributeTypeMap;
    }
}