// api/model/checkoutRequest.ts
/**
 * Ecommerce Demo API
 * Stable, codegen-oriented OpenAPI contract. Designed for multi-project, multi-language use (client/server stubs, DTOs, SDKs).
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CheckoutRequest {
    'email'?: string;
    /**
     * Optional order notes
     */
    'notes'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'email',
            baseName: 'email',
            type: 'string'
        },
        {
            name: 'notes',
            baseName: 'notes',
            type: 'string'
        }
    ];

    static getAttributeTypeMap() {
        return CheckoutRequest.attributeTypeMap;
    }
}