// api/model/checkoutResponse.ts
/**
 * Ecommerce Demo API
 * Stable, codegen-oriented OpenAPI contract. Designed for multi-project, multi-language use (client/server stubs, DTOs, SDKs).
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Order } from './order';

export class CheckoutResponse {
    'order': Order;
    'message'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'order',
            baseName: 'order',
            type: 'Order'
        },
        {
            name: 'message',
            baseName: 'message',
            type: 'string'
        }
    ];

    static getAttributeTypeMap() {
        return CheckoutResponse.attributeTypeMap;
    }
}