// api/model/deleteProductRequest.ts
/**
 * Ecommerce Demo API
 * Stable, codegen-oriented OpenAPI contract. Designed for multi-project, multi-language use (client/server stubs, DTOs, SDKs).
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class DeleteProductRequest {
    /**
     * Resource identifier
     */
    'id': string;
    'hardDelete'?: boolean = false;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'id',
            baseName: 'id',
            type: 'string'
        },
        {
            name: 'hardDelete',
            baseName: 'hardDelete',
            type: 'boolean'
        }
    ];

    static getAttributeTypeMap() {
        return DeleteProductRequest.attributeTypeMap;
    }
}